move labelled addresses to keystone as the only user
