+++ deployment/keystone/changeset/internal/state.go
@@ -12,8 +12,8 @@ import (
 	cldf_evm "github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
 	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 
-	"github.com/smartcontractkit/chainlink/deployment"
 	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
+	"github.com/smartcontractkit/chainlink/deployment/keystone/changeset/internal/addrbook"
 
 	capabilities_registry "github.com/smartcontractkit/chainlink-evm/gethwrappers/keystone/generated/capabilities_registry_1_1_0"
 	forwarder "github.com/smartcontractkit/chainlink-evm/gethwrappers/keystone/generated/forwarder_1_0_0"
@@ -74,7 +74,7 @@ func GetContractSets(lggr logger.Logger, req *GetContractSetsRequest) (*GetContr
 
 		// TODO: we need to expand/refactor the way labeled addresses are filtered
 		// see: https://smartcontract-it.atlassian.net/browse/CRE-363
-		filtered := deployment.LabeledAddresses(addrs).And(req.Labels...)
+		filtered := addrbook.LabeledAddresses(addrs).And(req.Labels...)
 
 		maps.Copy(filtered, forwarderAddrs)
